Created LoginButtonState.swift and LoginButtonViewModel.swift files to clearly seperate the code from LoginButton.swift
