+++ hindsight/Components/LoginButton/LoginButton.swift
@@ -10,82 +10,6 @@ import UIKit
 import SnapKit
 import RxSwift
 
-protocol LoginButtonStateProtocol {
-    /// The `Variable` represents the button text in the current state
-    var text: String {get}
-    /// The `UIColor` represents the button textColor in the current state
-    var textColor: UIColor {get}
-    /// The `UIColor` represents the button backgroundColor in the current state
-    var backgroundColor: UIColor {get}
-    /// The `UIImage` represents the button image in the current state
-    var image: UIImage? {get}
-}
-
-struct LoginButtonState: LoginButtonStateProtocol {
-    /// The `Variable` represents the button text in the current state
-    let text: String
-    /// The `UIColor` represents the button textColor in the current state
-    let textColor: UIColor
-    /// The `UIColor` represents the button backgroundColor in the current state
-    let backgroundColor: UIColor
-    /// The `UIImage` represents the button image in the current state
-    let image: UIImage?
-    /// The `Function` to construct the struct
-    init(text: String, textColor: UIColor, backgroundColor: UIColor, image: UIImage?) {
-        self.text = text
-        self.textColor = textColor
-        self.backgroundColor = backgroundColor
-        self.image = image
-    }
-}
-
-protocol LoginButtonProtocol {
-
-    var buttonState: Variable<LoginButtonStateProtocol> { get }
-
-    func didSelect()
-
-    func didCancel()
-
-    func didRelease()
-}
-
-struct LoginButtonViewModel: LoginButtonProtocol {
-
-    /// The `Variable` represents a RxSwift `Observable` which will be observed for button state changes
-    internal let buttonState: Variable<LoginButtonStateProtocol>
-    /// The `LoginButtonStateProtocol` represents button normal state
-    private let normalState: LoginButtonStateProtocol
-    /// The `LoginButtonStateProtocol` represents button highlight state
-    private let highlightState: LoginButtonStateProtocol?
-    /// The `VoidClosure` delegates the button tapped action
-    private let didSelectClosure: VoidClosure
-
-    init(normalState: LoginButtonStateProtocol,
-         highlightState: LoginButtonStateProtocol?,
-         didSelectClosure: @escaping VoidClosure) {
-        self.normalState = normalState
-        self.highlightState = highlightState
-        self.didSelectClosure = didSelectClosure
-        self.buttonState = Variable(normalState)
-    }
-
-    func didSelect() {
-        if let hs = highlightState {
-            buttonState.value = hs
-        }
-    }
-
-    func didCancel() {
-        buttonState.value = normalState
-    }
-
-    func didRelease() {
-        buttonState.value = normalState
-        didSelectClosure()
-    }
-}
-
 /// This UIView is created to match the design and can be reused in the future
 /// - Ex: Google Login, Instagram Login
 class LoginButton: UIView {
